requirements.txt
pyserial
numpy
matplotlib
scipy
pandas
RPi.GPIO
spidev
smbus
